# Faster bilinear resize for x86