--- a/tensorflow/core/kernels/crop_and_resize_op.cc
+++ b/tensorflow/core/kernels/crop_and_resize_op.cc
@@ -22,17 +22,19 @@ limitations under the License.
 #include <functional>
 #include <string>
 
-#include "third_party/eigen3/unsupported/Eigen/CXX11/Tensor"
 #include "tensorflow/core/framework/bounds_check.h"
 #include "tensorflow/core/framework/register_types.h"
 #include "tensorflow/core/framework/tensor.h"
 #include "tensorflow/core/framework/tensor_shape.h"
 #include "tensorflow/core/framework/types.h"
+#include "tensorflow/core/kernels/crop_resize_bilinear_core.h"
+#include "tensorflow/core/kernels/image_resizer_state.h"
 #include "tensorflow/core/lib/core/errors.h"
 #include "tensorflow/core/lib/core/status.h"
 #include "tensorflow/core/platform/logging.h"
 #include "tensorflow/core/platform/types.h"
 #include "tensorflow/core/util/work_sharder.h"
+#include "third_party/eigen3/unsupported/Eigen/CXX11/Tensor"
 
 #if GOOGLE_CUDA
 #include "tensorflow/core/common_runtime/gpu/gpu_event_mgr.h"
@@ -228,61 +230,58 @@ struct CropAndResize<CPUDevice, T> {
           continue;
         }
 
-        const float height_scale =
-            (crop_height > 1)
-                ? (y2 - y1) * (image_height - 1) / (crop_height - 1)
-                : 0;
-        const float width_scale =
-            (crop_width > 1) ? (x2 - x1) * (image_width - 1) / (crop_width - 1)
-                             : 0;
+        if (method_name == "bilinear") {
+          std::vector<CachedInterpolation> xs;
+          std::vector<CachedInterpolation> ys;
+          int min_ix, max_ix, min_iy, max_iy;
+          // TODO: Should the half_pixel_centers be added to this function?
+          compute_interpolation_weights(LegacyScaler(), crop_width, image_width,
+                                        x1, x2, &min_ix, &max_ix, &xs);
+          compute_interpolation_weights(LegacyScaler(), crop_height,
+                                        image_height, y1, y2, &min_iy, &max_iy,
+                                        &ys);
 
-        for (int y = 0; y < crop_height; ++y) {
-          const float in_y = (crop_height > 1)
-                                 ? y1 * (image_height - 1) + y * height_scale
-                                 : 0.5 * (y1 + y2) * (image_height - 1);
-          if (in_y < 0 || in_y > image_height - 1) {
-            for (int x = 0; x < crop_width; ++x) {
-              for (int d = 0; d < depth; ++d) {
-                crops(b, y, x, d) = extrapolation_value;
-              }
-            }
-            continue;
+          // multiply by depth to avoid multiplication in resize_single_image.
+          for (int i = min_ix; i <= max_ix; ++i) {
+            xs[i - min_ix].lower *= depth;
+            xs[i - min_ix].upper *= depth;
           }
-          if (method_name == "bilinear") {
-            const int top_y_index = floorf(in_y);
-            const int bottom_y_index = ceilf(in_y);
-            const float y_lerp = in_y - top_y_index;
 
-            for (int x = 0; x < crop_width; ++x) {
-              const float in_x = (crop_width > 1)
-                                     ? x1 * (image_width - 1) + x * width_scale
-                                     : 0.5 * (x1 + x2) * (image_width - 1);
-              if (in_x < 0 || in_x > image_width - 1) {
+          crop_resize_single_image_common<T, float>(
+              image.data() +
+                  static_cast<int64>(b_in) * static_cast<int64>(image_height) *
+                      static_cast<int64>(image_width) *
+                      static_cast<int64>(depth),
+              image_height, image_width, crop_height, crop_width, depth, min_ix,
+              max_ix, xs.data(), min_iy, max_iy, ys.data(), extrapolation_value,
+              false, false,
+              crops.data() +
+                  static_cast<int64>(b) * static_cast<int64>(crop_height) *
+                      static_cast<int64>(crop_width) *
+                      static_cast<int64>(depth));
+          // xs and ys are deallocated automatically when they go out of scope
+        } else {  // method == "nearest"
+          const float height_scale =
+              (crop_height > 1)
+                  ? (y2 - y1) * (image_height - 1) / (crop_height - 1)
+                  : 0;
+          const float width_scale =
+              (crop_width > 1)
+                  ? (x2 - x1) * (image_width - 1) / (crop_width - 1)
+                  : 0;
+
+          for (int y = 0; y < crop_height; ++y) {
+            const float in_y = (crop_height > 1)
+                                   ? y1 * (image_height - 1) + y * height_scale
+                                   : 0.5 * (y1 + y2) * (image_height - 1);
+            if (in_y < 0 || in_y > image_height - 1) {
+              for (int x = 0; x < crop_width; ++x) {
                 for (int d = 0; d < depth; ++d) {
                   crops(b, y, x, d) = extrapolation_value;
                 }
-                continue;
               }
-              const int left_x_index = floorf(in_x);
-              const int right_x_index = ceilf(in_x);
-              const float x_lerp = in_x - left_x_index;
-
-              for (int d = 0; d < depth; ++d) {
-                const float top_left(static_cast<float>(
-                    image(b_in, top_y_index, left_x_index, d)));
-                const float top_right(static_cast<float>(
-                    image(b_in, top_y_index, right_x_index, d)));
-                const float bottom_left(static_cast<float>(
-                    image(b_in, bottom_y_index, left_x_index, d)));
-                const float bottom_right(static_cast<float>(
-                    image(b_in, bottom_y_index, right_x_index, d)));
-                const float top = top_left + (top_right - top_left) * x_lerp;
-                const float bottom =
-                    bottom_left + (bottom_right - bottom_left) * x_lerp;
-                crops(b, y, x, d) = top + (bottom - top) * y_lerp;
-              }
+              continue;
             }
-          } else {  // method == "nearest"
             for (int x = 0; x < crop_width; ++x) {
               const float in_x = (crop_width > 1)
                                      ? x1 * (image_width - 1) + x * width_scale