--- a/tensorflow/core/kernels/resize_bilinear_op.cc
+++ b/tensorflow/core/kernels/resize_bilinear_op.cc
@@ -19,15 +19,16 @@ limitations under the License.
 #include "tensorflow/core/kernels/resize_bilinear_op.h"
 
 #include <memory>
-#include "third_party/eigen3/unsupported/Eigen/CXX11/Tensor"
 #include "tensorflow/core/framework/op_kernel.h"
 #include "tensorflow/core/framework/register_types.h"
 #include "tensorflow/core/framework/tensor.h"
 #include "tensorflow/core/framework/tensor_shape.h"
 #include "tensorflow/core/framework/types.h"
+#include "tensorflow/core/kernels/crop_resize_bilinear_core.h"
 #include "tensorflow/core/kernels/image_resizer_state.h"
 #include "tensorflow/core/lib/core/status.h"
 #include "tensorflow/core/platform/logging.h"
+#include "third_party/eigen3/unsupported/Eigen/CXX11/Tensor"
 
 namespace tensorflow {
 
@@ -66,145 +67,6 @@ class ResizeBilinearOp : public OpKernel {
   bool half_pixel_centers_;
 };
 
-namespace {
-// Compute the interpolation indices only once.
-struct CachedInterpolation {
-  int64 lower;  // Lower source index used in the interpolation
-  int64 upper;  // Upper source index used in the interpolation
-  // 1-D linear iterpolation scale (see:
-  // https://en.wikipedia.org/wiki/Bilinear_interpolation)
-  float lerp;
-};
-
-template <typename Scaler>
-inline void compute_interpolation_weights(const Scaler scaler,
-                                          const int64 out_size,
-                                          const int64 in_size,
-                                          const float scale,
-                                          CachedInterpolation* interpolation) {
-  interpolation[out_size].lower = 0;
-  interpolation[out_size].upper = 0;
-  for (int64 i = out_size - 1; i >= 0; --i) {
-    const float in = scaler(i, scale);
-    const float in_f = std::floor(in);
-    interpolation[i].lower =
-        std::max(static_cast<int64>(in_f), static_cast<int64>(0));
-    interpolation[i].upper =
-        std::min(static_cast<int64>(std::ceil(in)), in_size - 1);
-    interpolation[i].lerp = in - in_f;
-  }
-}
-
-/**
- * Computes the bilinear interpolation from the appropriate 4 float points
- * and the linear interpolation weights.
- */
-inline float compute_lerp(const float top_left, const float top_right,
-                          const float bottom_left, const float bottom_right,
-                          const float x_lerp, const float y_lerp) {
-  const float top = top_left + (top_right - top_left) * x_lerp;
-  const float bottom = bottom_left + (bottom_right - bottom_left) * x_lerp;
-  return top + (bottom - top) * y_lerp;
-}
-
-template <typename T>
-void resize_image(
-    typename TTypes<T, 4>::ConstTensor images, const int batch_size,
-    const int64 in_height, const int64 in_width, const int64 out_height,
-    const int64 out_width, const int channels,
-    const std::vector<CachedInterpolation>& xs,
-    const std::vector<CachedInterpolation>& ys,
-    typename TTypes<float, 4>::Tensor output) TF_ATTRIBUTE_NOINLINE;
-template <typename T>
-void resize_image(typename TTypes<T, 4>::ConstTensor images,
-                  const int batch_size, const int64 in_height,
-                  const int64 in_width, const int64 out_height,
-                  const int64 out_width, const int channels,
-                  const std::vector<CachedInterpolation>& xs_vec,
-                  const std::vector<CachedInterpolation>& ys,
-                  typename TTypes<float, 4>::Tensor output) {
-  const int64 in_row_size = in_width * channels;
-  const int64 in_batch_num_values = in_height * in_row_size;
-  const int64 out_row_size = out_width * channels;
-
-  const T* input_b_ptr = images.data();
-  const CachedInterpolation* xs = xs_vec.data();
-
-  if (channels == 3) {
-    float* output_y_ptr = output.data();
-    for (int b = 0; b < batch_size; ++b) {
-      for (int64 y = 0; y < out_height; ++y) {
-        const T* ys_input_lower_ptr = input_b_ptr + ys[y].lower * in_row_size;
-        const T* ys_input_upper_ptr = input_b_ptr + ys[y].upper * in_row_size;
-        const float ys_lerp = ys[y].lerp;
-        for (int64 x = 0; x < out_width; ++x) {
-          const int64 xs_lower = xs[x].lower;
-          const int64 xs_upper = xs[x].upper;
-          const float xs_lerp = xs[x].lerp;
-
-          // Read channel 0.
-          const float top_left0(ys_input_lower_ptr[xs_lower + 0]);
-          const float top_right0(ys_input_lower_ptr[xs_upper + 0]);
-          const float bottom_left0(ys_input_upper_ptr[xs_lower + 0]);
-          const float bottom_right0(ys_input_upper_ptr[xs_upper + 0]);
-
-          // Read channel 1.
-          const float top_left1(ys_input_lower_ptr[xs_lower + 1]);
-          const float top_right1(ys_input_lower_ptr[xs_upper + 1]);
-          const float bottom_left1(ys_input_upper_ptr[xs_lower + 1]);
-          const float bottom_right1(ys_input_upper_ptr[xs_upper + 1]);
-
-          // Read channel 2.
-          const float top_left2(ys_input_lower_ptr[xs_lower + 2]);
-          const float top_right2(ys_input_lower_ptr[xs_upper + 2]);
-          const float bottom_left2(ys_input_upper_ptr[xs_lower + 2]);
-          const float bottom_right2(ys_input_upper_ptr[xs_upper + 2]);
-
-          // Compute output.
-          output_y_ptr[x * channels + 0] =
-              compute_lerp(top_left0, top_right0, bottom_left0, bottom_right0,
-                           xs_lerp, ys_lerp);
-          output_y_ptr[x * channels + 1] =
-              compute_lerp(top_left1, top_right1, bottom_left1, bottom_right1,
-                           xs_lerp, ys_lerp);
-          output_y_ptr[x * channels + 2] =
-              compute_lerp(top_left2, top_right2, bottom_left2, bottom_right2,
-                           xs_lerp, ys_lerp);
-        }
-        output_y_ptr += out_row_size;
-      }
-      input_b_ptr += in_batch_num_values;
-    }
-  } else {
-    float* output_y_ptr = output.data();
-    for (int b = 0; b < batch_size; ++b) {
-      for (int64 y = 0; y < out_height; ++y) {
-        const T* ys_input_lower_ptr = input_b_ptr + ys[y].lower * in_row_size;
-        const T* ys_input_upper_ptr = input_b_ptr + ys[y].upper * in_row_size;
-        const float ys_lerp = ys[y].lerp;
-        for (int64 x = 0; x < out_width; ++x) {
-          auto xs_lower = xs[x].lower;
-          auto xs_upper = xs[x].upper;
-          auto xs_lerp = xs[x].lerp;
-          for (int c = 0; c < channels; ++c) {
-            const float top_left(ys_input_lower_ptr[xs_lower + c]);
-            const float top_right(ys_input_lower_ptr[xs_upper + c]);
-            const float bottom_left(ys_input_upper_ptr[xs_lower + c]);
-            const float bottom_right(ys_input_upper_ptr[xs_upper + c]);
-            output_y_ptr[x * channels + c] =
-                compute_lerp(top_left, top_right, bottom_left, bottom_right,
-                             xs_lerp, ys_lerp);
-          }
-        }
-        output_y_ptr += out_row_size;
-      }
-      input_b_ptr += in_batch_num_values;
-    }
-  }
-}
-
-}  // namespace
-
 // Partial specialization of ResizeBilinear functor for a CPUDevice.
 namespace functor {
 template <typename T>
@@ -221,14 +83,23 @@ struct ResizeBilinear<CPUDevice, T> {
     const int64 out_height = output.dimension(1);
     const int64 out_width = output.dimension(2);
 
+    const int64 in_row_size = in_width * channels;
+    const int64 in_batch_num_values = in_height * in_row_size;
+    const int64 out_row_size = out_width * channels;
+    const int64 out_batch_num_values = out_row_size * out_height;
+
     // Handle no-op resizes efficiently.
     if (out_height == in_height && out_width == in_width) {
       output = images.template cast<float>();
       return;
     }
 
-    std::vector<CachedInterpolation> ys(out_height + 1);
-    std::vector<CachedInterpolation> xs(out_width + 1);
+    // Compute the cached interpolation weights on the x and y dimensions.
+    std::vector<CachedInterpolation> ys;
+    ys.resize(out_height + 1);
+
+    std::vector<CachedInterpolation> xs;
+    xs.resize(out_width + 1);
 
     if (half_pixel_centers) {
       compute_interpolation_weights(HalfPixelScaler(), out_height, in_height,
@@ -243,14 +114,21 @@ struct ResizeBilinear<CPUDevice, T> {
       compute_interpolation_weights(LegacyScaler(), out_width, in_width,
                                     width_scale, xs.data());
     }
+
     // Scale x interpolation weights to avoid a multiplication during iteration.
     for (int i = 0; i < xs.size(); ++i) {
       xs[i].lower *= channels;
       xs[i].upper *= channels;
     }
 
-    resize_image<T>(images, batch_size, in_height, in_width, out_height,
-                    out_width, channels, xs, ys, output);
+    for (int b = 0; b < batch_size; ++b) {
+      crop_resize_single_image_common(
+          images.data() + (int64)b * in_batch_num_values, in_height, in_width,
+          out_height, out_width, channels, 0, out_width - 1, xs.data(), 0,
+          out_height - 1, ys.data(), 0.0f, false, false,
+          output.data() + (int64)b * out_batch_num_values);
+    }
+    // xs and ys are freed when they go out of scope
   }
 };
 }  // namespace functor