… tensorflow/core/kernels/crop_resize_bilinear_core.h