- Resubmission of PR 21681
- Added F16C check around _mm_cvtph_ps function
- Integrated changes with half-pixel centers