@MattConley 2019-04-10
Thanks for the review, I'm making progress with the changes and will push the fixes soon.