@rmlarsen 2019-04-23
This code is pretty unreadable. @martinwicke  do you have any guidance? I don't think we can accept this PR as is. @MattConley  Could you reuse parts of Eigen instead of adding your own library of low-level operations using intrinsics directly?