@martinwicke 2019-04-23
Hmm... I don't see where intrinsics are used directly here? If you cannot read the code that's a bad sign though.