@penpornk 2019-04-24
@martinwicke Intrinsics are in `crop_resize_bilinear_core.h`. It is pretty big so github collapses the file by default.