@rmlarsen 2019-04-24
@martinwicke I exaggerated a bit, but it basically looks like a lot of plain C code.