@tfboyd 2019-05-01
NVIDIA will do an A/B test and determine if it is worth reworking this in Eigan or just closing.