@MattConley 2019-05-03
Here are the requested performance numbers run on my local machine with an Intel i7-5820K

Without the changes:
```
Benchmark                                           Time(ns) Iterations
-----------------------------------------------------------------------
BM_CropAndResize_cpu_DT_UINT8_1_640_640_3_512_512    6270880        100	 196.0M items/s
BM_CropAndResize_cpu_DT_UINT8_1_640_640_1_512_512    3954793        164	 103.6M items/s
BM_CropAndResize_cpu_DT_HALF_1_640_640_3_512_512     9292230        100	 132.2M items/s
BM_CropAndResize_cpu_DT_HALF_1_640_640_1_512_512     5378040        100	 76.2M items/s
BM_CropAndResize_cpu_DT_FLOAT_1_640_640_3_512_512    5738020        100	 214.2M items/s
BM_CropAndResize_cpu_DT_FLOAT_1_640_640_1_512_512    5039507        150	 81.3M items/s
BM_CropAndResize_cpu_DT_FLOAT_1_80_80_512_7_7         140624       4096	 932.1M items/s
```

With the changes:
```
Benchmark                                           Time(ns) Iterations
-----------------------------------------------------------------------
BM_CropAndResize_cpu_DT_UINT8_1_640_640_3_512_512    3932621        198  312.5M items/s 
BM_CropAndResize_cpu_DT_UINT8_1_640_640_1_512_512    1568988        427  261.1M items/s 
BM_CropAndResize_cpu_DT_HALF_1_640_640_3_512_512     6684210        100  183.8M items/s 
BM_CropAndResize_cpu_DT_HALF_1_640_640_1_512_512     2892926        204  141.6M items/s 
BM_CropAndResize_cpu_DT_FLOAT_1_640_640_3_512_512    2587333        267  474.9M items/s 
BM_CropAndResize_cpu_DT_FLOAT_1_640_640_1_512_512    1094434        585  374.3M items/s 
BM_CropAndResize_cpu_DT_FLOAT_1_80_80_512_7_7         132819       4054  741.5M items/s 
```