@tfboyd 2019-05-29
NVIDIA is closing based on not having internal experience with Eigen.  Would need to be picked up by TesnorFlow and maybe team up if needed.